requirements.txt
streamlit==1.32.0
pdfplumber==0.10.4
python-docx==1.1.0
nltk==3.8.1
scikit-learn==1.4.1.post1
pandas==2.2.1
plotly==5.20.0
Pillow==10.2.0
reportlab==4.1.0
